login: assert server provided gait_path is recoverable
At registration, the client uploads to the server the "gait path",
which is used to the derive the server keys for the wallet. This
value is normally derived using the "gait xpub" (M/0x4741'), but
some older clients did this slightly differently. These different
derivations are supported by the recovery tool "garecovery".

At each login, the server provides the wallet with the previously
uploaded "gait path". However, if the server provided an incorrect
"gait path", the client would have used it without realizing it,
thus loosing its ability to recover funds in case the server
becomes unavailable. This commit prevents that from happening.

Note: this change implies that wallets with a "gait path" which is
not supported by "garecovery" won't be able to login. In addition,
wallets which both derived the "gait path" directly from the mnemonic
(rather than an extended public key) and are now using an
hardware wallet, won't be able to login too.